Create SetLobbyPlayedWith() method
I have a theory that I might want to call this on a timer... I'll have to test it, because the docs for when to call SetPlayedWith simply don't exist. The docs tell you what it does, but not best practices for when to call it, or if you should call it multiple times.